Add BasePrimitive class
